+++ src/components/Link.react.js
@@ -4,6 +4,8 @@ import PropTypes from 'prop-types';
 
 import React, {Component} from 'react';
 
+import {isNil} from 'ramda';
+
 /*
  * event polyfill for IE
  * https://developer.mozilla.org/en-US/docs/Web/API/CustomEvent/CustomEvent
@@ -54,7 +56,15 @@ export default class Link extends Component {
     }
 
     render() {
-        const {className, style, id, href, loading_state} = this.props;
+        const {
+            className,
+            style,
+            id,
+            href,
+            loading_state,
+            children,
+            title,
+        } = this.props;
         /*
          * ideally, we would use cloneElement however
          * that doesn't work with dash's recursive
@@ -70,8 +80,9 @@ export default class Link extends Component {
                 style={style}
                 href={href}
                 onClick={e => this.updateLocation(e)}
+                title={title}
             >
-                {this.props.children}
+                {isNil(children) ? href : children}
             </a>
         );
     }
@@ -87,7 +98,7 @@ Link.propTypes = {
     /**
      * The URL of a linked resource.
      */
-    href: PropTypes.string,
+    href: PropTypes.string.isRequired,
     /**
      * Controls whether or not the page will refresh when the link is clicked
      */
@@ -100,6 +111,11 @@ Link.propTypes = {
      * Defines CSS styles which will override styles previously set.
      */
     style: PropTypes.object,
+    /**
+     * Adds the title attribute to your link, which can contain supplementary
+     * information.
+     */
+    title: PropTypes.string,
     /**
      * The children of this component
      */